# public/_doc/swagger.yaml
swagger: '2.0'
info:
  description: OpenAI API ドキュメント
  version: 0.0.1
  title: OpenAI API REST APIs
#host: urapi.herokuapp.com
basePath: /api
tags:
  - name: openai
    description: OpenAI API
schemes:
  - http
  - https
paths:
  /ping:
    get:
      tags:
        - openai
      summary: ping
      description: ping(test)
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /models:
    get:
      tags:
        - openai
      summary: Models API
      description: Retrieve Models
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /model/{id}:
    get:
      tags:
        - openai
      summary: Retrieve Model API
      description: Retrieve Model
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /excel:
    post:
      tags:
        - openai
      summary: Analyse with Excel file
      description: Analyse with Excel file
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: prompt
          in: formData
          type: string
          description: 指示プロンプト
        - name: file
          in: formData
          type: file
          description: xlsx ファイル
          required: true
        - name: model
          in: formData
          type: string
          enum: [ "gpt-3.5-turbo", "gpt-4o-mini", "gpt-4" ]
          description: モデルID
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
        '429':
          description: Too many requests
  /csv:
    post:
      tags:
        - openai
      summary: Analyse with CSV file
      description: Analyse with CSV file
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: prompt
          in: formData
          type: string
          description: 指示プロンプト
        - name: file
          in: formData
          type: file
          description: csv ファイル
          required: true
        - name: model
          in: formData
          type: string
          enum: [ "gpt-3.5-turbo", "gpt-4o-mini", "gpt-4" ]
          description: モデルID
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
        '429':
          description: Too many requests